main_service/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
gunicorn==22.0.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
redis[hiredis]==5.0.4
python-multipart==0.0.9
httpx==0.27.0
google-generativeai==0.5.4
google-cloud-storage==2.16.0
pydantic==2.7.1
pydantic-settings==2.2.1
alembic==1.13.1
pytest==8.2.0
pytest-asyncio==0.23.6
